Fix incorrect usage of std::set_intersection
std::set_intersection is only defined for std::set, not for an unordered
set. This change fixes all wrong usages in the overlap function by
replacing them with a for loop.

The wrong usage had no effect on correctness, since overlap() is only
used in DEBUG mode for an assertion check.

Fixes issue #18.